Cleaning onCallModifying
This function is responsible for building the 200 OK to answer the REINVITE coming from Freeswitch. In our cause, we just need to use bbb-voice local sdp to answer it. To a later and permanent change, however, it's recommended to investigate the SdpTools.java. My guess is that sdpMediaProduct function is comparing the media codecs in a wrong way.